Move context managers to a file of their own
