+++ prometheus_client/core.py
@@ -11,11 +11,9 @@ import time
 import types
 
 from threading import Lock
-from timeit import default_timer
 from collections import namedtuple
 
-from .decorator import decorate
-
+from .context_managers import _ExceptionCounter, _InprogressTracker, _Timer
 
 if sys.version_info > (3,):
     unicode = str
@@ -1007,61 +1005,3 @@ class Enum(object):
                        for i, s in enumerate(self._states)]
 
 
-class _ExceptionCounter(object):
-    def __init__(self, counter, exception):
-        self._counter = counter
-        self._exception = exception
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, typ, value, traceback):
-        if isinstance(value, self._exception):
-            self._counter.inc()
-
-    def __call__(self, f):
-        def wrapped(func, *args, **kwargs):
-            with self:
-                return func(*args, **kwargs)
-        return decorate(f, wrapped)
-
-
-class _InprogressTracker(object):
-    def __init__(self, gauge):
-        self._gauge = gauge
-
-    def __enter__(self):
-        self._gauge.inc()
-
-    def __exit__(self, typ, value, traceback):
-        self._gauge.dec()
-
-    def __call__(self, f):
-        def wrapped(func, *args, **kwargs):
-            with self:
-                return func(*args, **kwargs)
-        return decorate(f, wrapped)
-
-
-class _Timer(object):
-    def __init__(self, callback):
-        self._callback = callback
-
-    def _new_timer(self):
-        return self.__class__(self._callback)
-
-    def __enter__(self):
-        self._start = default_timer()
-
-    def __exit__(self, typ, value, traceback):
-        # Time can go backwards.
-        duration = max(default_timer() - self._start, 0)
-        self._callback(duration)
-
-    def __call__(self, f):
-        def wrapped(func, *args, **kwargs):
-            # Obtaining new instance of timer every time
-            # ensures thread safety and reentrancy.
-            with self._new_timer():
-                return func(*args, **kwargs)
-        return decorate(f, wrapped)
